+++ src/conferencehall/components/ConferenceHallProposalsPicker.tsx
@@ -6,7 +6,7 @@ import { Box, Chip, Typography } from '@mui/material'
 import LoadingButton from '@mui/lab/LoadingButton'
 import { DataGrid, GridColDef, GridToolbarQuickFilter, useGridApiContext } from '@mui/x-data-grid'
 import { DocumentData } from '@firebase/firestore'
-import { UseQueryResult } from '../../services/hooks/firebaseQueryHook'
+import { UseQueryResult } from '../../services/hooks/firestoreQueryHook'
 
 const columns: GridColDef[] = [
     { field: 'title', headerName: 'Title', width: 350 },
