+++ src/conferencehall/hooks/useConferenceHallEvents.tsx
@@ -1,7 +1,7 @@
 import { DocumentData, query, QueryConstraint } from '@firebase/firestore'
 import { where } from 'firebase/firestore'
 import { conferenceHallCollections } from '../firebase/conferenceHallFirebase'
-import { useFirestoreCollection, UseQueryResult } from '../../services/hooks/firebaseQueryHook'
+import { useFirestoreCollection, UseQueryResult } from '../../services/hooks/firestoreQueryHook'
 
 export const useConferenceHallEvents = (userId: string): UseQueryResult<DocumentData> => {
     const constraints: QueryConstraint[] = []
