+++ src/conferencehall/hooks/useFakeConferenceHallProposals.ts
@@ -3,7 +3,7 @@ import { useEffect, useState } from 'react'
 import { v4 as uuidv4 } from 'uuid'
 import { randomEnum } from '../../utils/getRandomEnum'
 import { DocumentData } from '@firebase/firestore'
-import { UseQueryResult } from '../../services/hooks/firebaseQueryHook'
+import { UseQueryResult } from '../../services/hooks/firestoreQueryHook'
 
 export const useFakeConferenceHallProposals = (eventId: string, count = 150): UseQueryResult<DocumentData> => {
     const [proposals, setProposals] = useState<ConferenceHallProposal[]>([])
