+++ src/services/hooks/useEvent.ts
@@ -1,6 +1,6 @@
 import { collections } from '../firebase'
 import { doc, DocumentData } from '@firebase/firestore'
-import { useFirestoreDocument, UseQueryResult } from './firebaseQueryHook'
+import { useFirestoreDocument, UseQueryResult } from './firestoreQueryHook'
 
 export const useEvent = (eventId?: string): UseQueryResult<DocumentData> => {
     const ref = doc(collections.events, eventId)
