+++ src/services/hooks/useEvents.ts
@@ -2,7 +2,7 @@ import { collections } from '../firebase'
 import { DocumentData, query, QueryConstraint } from '@firebase/firestore'
 import { where } from 'firebase/firestore'
 import { UserId } from '../../auth/authReducer'
-import { useFirestoreCollection, UseQueryResult } from './firebaseQueryHook'
+import { useFirestoreCollection, UseQueryResult } from './firestoreQueryHook'
 
 export const useEvents = (userId: UserId): UseQueryResult<DocumentData> => {
     const constraints: QueryConstraint[] = []
