+++ src/services/hooks/useSpeaker.ts
@@ -1,7 +1,7 @@
 import { collections } from '../firebase'
 import { doc } from 'firebase/firestore'
 import { Speaker } from '../../types'
-import { useFirestoreDocument, UseQueryResult } from './firebaseQueryHook'
+import { useFirestoreDocument, UseQueryResult } from './firestoreQueryHook'
 
 export const useSpeaker = (eventId: string, speakerId: string): UseQueryResult<Speaker> => {
     return useFirestoreDocument(doc(collections.speakers(eventId), speakerId), true)
