+++ src/services/hooks/useSponsors.ts
@@ -1,6 +1,6 @@
 import { collections } from '../firebase'
 import { DocumentData, query } from '@firebase/firestore'
-import { useFirestoreCollection, UseQueryResult } from './firebaseQueryHook'
+import { useFirestoreCollection, UseQueryResult } from './firestoreQueryHook'
 
 export const useSponsors = (eventId: string): UseQueryResult<DocumentData> => {
     return useFirestoreCollection(query(collections.sponsors(eventId)), true)
